+++ src/dms-patterns/core/table-mappings/index.ts
@@ -2,7 +2,8 @@
 All Rights Reserved.
 */
 
+export * from './base';
 export * from './s3';
 export * from './selection';
-export * from './transformation';
 export * from './table-settings';
+export * from './transformation';
