+++ src/dms-patterns/core/table-mappings/s3.ts
